Time: 186 ms (29.71%), Space: 107.6 MB (54.45%) - LeetHub

+++ 2348-number-of-zero-filled-subarrays/2348-number-of-zero-filled-subarrays.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
-        long res=0;
+        long long res=0;
         for(int i=0,j=0;i<nums.size();i++){
             if(nums[i]!=0)
                 j=i+1;
